move validate_request from decorators to openapi_tools (add back-ref)
