+++ rest_tools/server/decorators.py
@@ -3,7 +3,6 @@
 # fmt:off
 
 import logging
-import os
 import re
 from functools import wraps
 from inspect import isawaitable
@@ -12,7 +11,7 @@ from typing import Protocol
 import requests.exceptions
 import tornado.web
 
-from .. import telemetry as wtt
+from .. import openapi_tools, telemetry as wtt
 
 LOGGER = logging.getLogger(__name__)
 
@@ -383,96 +382,6 @@ class TokenAttributeRoleMappingProtocol(Protocol):
 
 
 ########################################################################################################################
-# fmt:on
-
-try:
-    import openapi_core
-    from openapi_core import OpenAPI
-    from openapi_core.contrib import requests as openapi_core_requests
-    from openapi_core.validation.exceptions import ValidationError
-    import openapi_core.validation.schemas.exceptions
-
-    openapi_available = True
-except ImportError:
-    # if client code wants to use these features, then let the built-in errors raise
-    openapi_available = False
 
 
-def validate_request(openapi_spec: "OpenAPI"):  # type: ignore
-    """Validate request obj against the given OpenAPI spec."""
-    if not openapi_available:
-        raise RuntimeError(
-            "openapi cannot be imported! perhaps you meant to pip install it?"
-        )
-
-    def make_wrapper(method):  # type: ignore[no-untyped-def]
-        async def wrapper(zelf: tornado.web.RequestHandler, *args, **kwargs):  # type: ignore[no-untyped-def]
-            LOGGER.debug("validating with openapi spec")
-            # NOTE - don't change data (unmarshal) b/c we are downstream of data separation
-            try:
-                # https://openapi-core.readthedocs.io/en/latest/validation.html
-                openapi_spec.validate_request(
-                    _http_server_request_to_openapi_request(zelf.request),
-                )
-            except ValidationError as e:  # type: ignore
-                LOGGER.error(f"invalid request: {e.__class__.__name__} - {e}")
-                if isinstance(  # look at the ORIGINAL exception that caused this error
-                    e.__context__,
-                    openapi_core.validation.schemas.exceptions.InvalidSchemaValue,  # type: ignore
-                ):
-                    reason = "; ".join(  # to client
-                        # verbose details after newline
-                        str(x).split("\n", maxsplit=1)[0]
-                        for x in e.__context__.schema_errors
-                    )
-                else:
-                    reason = str(e)  # to client
-                if os.getenv("CI"):
-                    # in prod, don't fill up logs w/ traces from invalid data
-                    LOGGER.exception(e)
-                raise tornado.web.HTTPError(
-                    status_code=400,
-                    log_message=f"{e.__class__.__name__}: {e}",  # to stderr
-                    reason=reason,  # to client
-                )
-            except Exception as e:
-                LOGGER.error(f"unexpected exception: {e.__class__.__name__} - {e}")
-                LOGGER.exception(e)
-                raise tornado.web.HTTPError(
-                    status_code=400,
-                    log_message=f"{e.__class__.__name__}: {e}",  # to stderr
-                    reason=None,  # to client (don't send possibly sensitive info)
-                )
-
-            return await method(zelf, *args, **kwargs)
-
-        return wrapper
-
-    return make_wrapper
-
-
-def _http_server_request_to_openapi_request(
-    req: tornado.httputil.HTTPServerRequest,
-) -> "openapi_core_requests.RequestsOpenAPIRequest":
-    """Convert a `tornado.httputil.HTTPServerRequest` to openapi's type."""
-    if not openapi_available:
-        raise RuntimeError(
-            "openapi cannot be imported! perhaps you meant to pip install it?"
-        )
-    return openapi_core_requests.RequestsOpenAPIRequest(  # type: ignore
-        requests.Request(
-            method=req.method.lower() if req.method else "get",
-            url=f"{req.protocol}://{req.host}{req.uri}",
-            headers=req.headers,
-            files=req.files,
-            data=req.body if not req.body_arguments else None,  # see below
-            params=req.query_arguments,
-            # auth=None,
-            cookies=req.cookies,
-            # hooks=None,
-            json=req.body_arguments if req.body_arguments else None,  # see above
-        )
-    )
-
-
-########################################################################################################################
+validate_request = openapi_tools.validate_request
